refactor : S3, Tmap 서비스 디렉터리 구조 변경
